Application: Move handling of dpi changes to it
